seperate interface, implementation, mock code
